requirements.txt
pyopenssl
pymemcache
beanstalkc3
sqlalchemy